make: Expand everything before running babel
Due to a setting in .orgconfig we only show the overview, which is
useful in interactive use.  We have to explicitly expand everything
before running babel because otherwise links are not propertized to
hide the parts that are supposed to be hidden, and because links
appear inside tables, that would cause those tables to be misaligned.